Drop departures ONLY if exists

+++ db/migrate/20151122024031_create_departures.rb
@@ -1,6 +1,6 @@
 class CreateDepartures < ActiveRecord::Migration
   def change
-    drop_table :departures
+    drop_table :departures if (table_exists? :departures)
     create_table :departures do |t|
       t.references :location, index: true, foreign_key: true
       t.references :trackable, index: true, foreign_key: true
